sched-ext: add scx_scheduler for multi scheduler support
Change the scx_has_op from static_key into bitmap. static_key is not
available for dynamic allocation of a struct, which is needed for multi
schedulers to be allocated on runtime. Further, bitmap is compact and
not time-consuming compared to bool array, which is 32x larger than
bitmap, consumes more cache space and only one less instruction.

The origin usage of static_key also alleviate the time advantage with
many places using static_key logically and/or other conditional
branches. Branch prediction is still needed in such cases.