sched-ext: add scx_scheduler for multi scheduler support (9)
Complete the priority mechanism: newest scheduler with one or more tasks
has highest priority, which is inserted as the first node of prio list.
And no task scheduler has lowest priority, which is inserted as the last
node of prio list. (By default the scheduler is empty, thus use
list_add_tail to insert the scheduler node.)