sched-ext: add scx_scheduler for multi scheduler support (13)
Fix the bug of reading _curr_sched percpu variable. Instead of
this_cpu_ptr, this_cpu_read is the correct choice to read the data of
the percpu variable. While using this_cpu_ptr, it will treats the input
value as an address which needs a base address added to be accessed. The
value of _curr_sched itself is an address of the corresponding
scx_scheduler, but it should not add the this_cpu_off base address.
this_cpu_read is reading _curr_sched directly, which is visible for
every CPU with its own memory address.