sched-ext: add scx_scheduler for multi scheduler support (15)
Fixed the issue of root_cgroup_path copying from uops to ops. uops is
not stored in the userspace, its data is stored in the kernel space as
well, which is copied from user-defined struct_ops by BPF mechanism
earlier.
Given this situation, `copy_from_user` is not appropriate here. Use
strscpy instead, which is immune to stack overflow attack.