sched-ext: add scx_scheduler for multi scheduler support (18)
Fixed tg->sched is NULL pointer problem, this is primarily because
`dummy_sched` doesn't correctly do the `scx_cgroup_init` for
`root_task_grp`.

Adding `scx_cgroup_init` in `scx_init` is not enough, because there are
possibilities to enable a `task_group` randomly after `scx_init`, which
is not the `task_group` for any user-defined schedulers. When this
happens, the `task_group` should have its parent's `task_group`.